requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
anthropic==0.84.0
sentence-transformers==3.1.1
PyMuPDF==1.24.10
python-docx==1.1.2
numpy==1.26.4
scikit-learn==1.5.2
aiofiles==24.1.0
authlib==1.3.2
httpx==0.27.2
itsdangerous==2.2.0
cryptography==43.0.1
starlette[full]==0.38.6
